feat(chat): move app rail to persistent sidebar so it sticks across routes
