+++ app/src/pages/Accounts.tsx
@@ -1,9 +1,6 @@
 import { useEffect, useMemo, useState } from 'react';
 
-import AddAccountModal from '../components/accounts/AddAccountModal';
-import { AgentIcon, ProviderIcon } from '../components/accounts/providerIcons';
 import WebviewHost from '../components/accounts/WebviewHost';
-import { SidebarContent } from '../components/layout/shell/SidebarSlot';
 import {
   CustomGifMascot,
   getMascotPalette,
@@ -13,101 +10,24 @@ import {
 import { useHumanMascot } from '../features/human/useHumanMascot';
 import { usePrewarmMostRecentAccount } from '../hooks/usePrewarmMostRecentAccount';
 import { useT } from '../lib/i18n/I18nContext';
-import { trackEvent } from '../services/analytics';
 import {
   hideWebviewAccount,
-  purgeWebviewAccount,
   showWebviewAccount,
   startWebviewAccountService,
 } from '../services/webviewAccountService';
-import {
-  addAccount,
-  removeAccount,
-  setActiveAccount,
-  setLastActiveAccount,
-} from '../store/accountsSlice';
-import { useAppDispatch, useAppSelector } from '../store/hooks';
+import { useAppSelector } from '../store/hooks';
 import {
   selectCustomMascotGifUrl,
   selectCustomPrimaryColor,
   selectCustomSecondaryColor,
   selectMascotColor,
 } from '../store/mascotSlice';
-import type { Account, AccountProvider, ProviderDescriptor } from '../types/accounts';
+import type { Account } from '../types/accounts';
 import { AGENT_ACCOUNT_ID as AGENT_ID } from '../utils/accountsFullscreen';
 import Conversations, { AgentChatPanel } from './Conversations';
 
 // Persistence key for face-toggle state across sessions.
 const FACE_MODE_KEY = 'chat.faceMode';
-
-function makeAccountId(): string {
-  const c = globalThis.crypto;
-  if (c && typeof c.randomUUID === 'function') return c.randomUUID();
-  if (c && typeof c.getRandomValues === 'function') {
-    const bytes = new Uint8Array(4);
-    c.getRandomValues(bytes);
-    const suffix = Array.from(bytes, b => b.toString(16).padStart(2, '0')).join('');
-    return `acct-${Date.now().toString(36)}-${suffix}`;
-  }
-  return `acct-${Date.now().toString(36)}-${Math.random().toString(36).slice(2, 8)}`;
-}
-
-interface RailButtonProps {
-  active: boolean;
-  onClick: () => void;
-  onContextMenu?: (e: React.MouseEvent) => void;
-  tooltip: string;
-  analyticsId: string;
-  badge?: number;
-  children: React.ReactNode;
-}
-
-const RailButton = ({
-  active,
-  onClick,
-  onContextMenu,
-  tooltip,
-  analyticsId,
-  badge,
-  children,
-}: RailButtonProps) => (
-  <button
-    type="button"
-    onClick={onClick}
-    onContextMenu={onContextMenu}
-    title={tooltip}
-    data-analytics-id={analyticsId}
-    // Issue #1284 — `hover:z-50` lifts the entire button (and its tooltip
-    // child) above sibling rail buttons during hover. Without it, the
-    // `hover:scale-105` transform on a non-active button establishes its
-    // own stacking context that traps the tooltip's `z-50` inside it,
-    // and a later sibling button (next in DOM order) paints over the
-    // tooltip rectangle. Belt-and-suspenders for the active-button case
-    // too, where ring-2 + bg-primary-50 don't transform but the lifted
-    // z still helps tooltips render cleanly above neighbours.
-    className={`group relative flex h-9 w-9 flex-none items-center justify-center rounded-lg transition-all hover:z-50 ${
-      active
-        ? 'bg-primary-50 ring-2 ring-primary-500'
-        : 'hover:bg-stone-100 dark:hover:bg-neutral-800/60 hover:scale-105'
-    }`}
-    aria-label={tooltip}>
-    {children}
-    {badge && badge > 0 ? (
-      <span className="absolute -right-0.5 -top-0.5 flex min-w-[16px] items-center justify-center rounded-full bg-coral-500 px-1 text-[9px] font-semibold text-white">
-        {badge > 99 ? '99+' : badge}
-      </span>
-    ) : null}
-    {/* Tooltip is the native `title` (set above): a custom absolute tooltip got
-        clipped by the rail's horizontal-scroll overflow, and the native title
-        isn't subject to overflow clipping or CEF webview compositing. */}
-  </button>
-);
-
-interface ContextMenuState {
-  accountId: string;
-  x: number;
-  y: number;
-}
 
 /**
  * Mascot + TTS panel rendered in face mode (right column of the Assistant
@@ -188,13 +108,12 @@ const FaceModePanel = () => {
 
 const Accounts = () => {
   const { t } = useT();
-  const dispatch = useAppDispatch();
   const accountsById = useAppSelector(state => state.accounts.accounts);
   const order = useAppSelector(state => state.accounts.order);
   const activeAccountId = useAppSelector(state => state.accounts.activeAccountId);
-  const unreadByAccount = useAppSelector(state => state.accounts.unread);
-  const [addOpen, setAddOpen] = useState(false);
-  const [ctxMenu, setCtxMenu] = useState<ContextMenuState | null>(null);
+  // Overlay state is owned by the persistent app rail (now in the sidebar); we
+  // only read it here to hide/restore the active provider webview.
+  const overlayOpen = useAppSelector(state => state.accounts.overlayOpen);
 
   const [faceMode] = useState<boolean>(() => {
     try {
@@ -222,22 +141,16 @@ const Accounts = () => {
   );
   usePrewarmMostRecentAccount({ accounts, accountsById, activeAccountId });
 
-  const connectedProviders = useMemo(
-    () => new Set<AccountProvider>(accounts.map(a => a.provider)),
-    [accounts]
-  );
-
   const selectedId = activeAccountId ?? AGENT_ID;
   const active = selectedId === AGENT_ID ? null : (accountsById[selectedId] ?? null);
   const isAgentSelected = selectedId === AGENT_ID;
 
   // The child Tauri webview is a native view composited above the HTML
   // canvas, so DOM z-index can't put React overlays on top of it. Hide
-  // the active webview while any overlay (add-account modal or the
-  // right-click context menu) is open and restore it on close. No-op
-  // when the agent pane is selected (pure HTML).
+  // the active webview while a rail overlay (add-account modal or the
+  // right-click context menu, both owned by the persistent sidebar rail) is
+  // open and restore it on close. No-op when the agent pane is selected.
   const activeId = active?.id ?? null;
-  const overlayOpen = addOpen || ctxMenu !== null;
   useEffect(() => {
     if (!activeId) return;
     if (overlayOpen) {
@@ -247,149 +160,12 @@ const Accounts = () => {
     }
   }, [overlayOpen, activeId]);
 
-  const handlePickProvider = (p: ProviderDescriptor) => {
-    setAddOpen(false);
-    trackEvent('account_connect_start', { provider: p.id });
-    const id = makeAccountId();
-    const acct: Account = {
-      id,
-      provider: p.id,
-      label: p.label,
-      createdAt: new Date().toISOString(),
-      status: 'pending',
-    };
-    dispatch(addAccount(acct));
-    dispatch(setActiveAccount(id));
-    // Issue #1233 — record this real-account selection in the persisted
-    // MRU pointer so the next session can prewarm it. Agent selections
-    // never reach this code path (separate `selectAgent` callback below).
-    dispatch(setLastActiveAccount(id));
-  };
-
-  const selectAgent = () => {
-    trackEvent('tauri_browser_click', {
-      surface: 'chat_right_sidebar',
-      action: 'select_agent',
-      provider: 'agent',
-    });
-    dispatch(setActiveAccount(AGENT_ID));
-  };
-  const selectAccount = (id: string) => {
-    const account = accountsById[id];
-    if (account) {
-      trackEvent('tauri_browser_click', {
-        surface: 'chat_right_sidebar',
-        action: 'select_account',
-        provider: account.provider,
-        account_status: account.status ?? 'unknown',
-      });
-    }
-    dispatch(setActiveAccount(id));
-    dispatch(setLastActiveAccount(id));
-  };
-
-  const openContextMenu = (accountId: string, e: React.MouseEvent) => {
-    e.preventDefault();
-    setCtxMenu({ accountId, x: e.clientX, y: e.clientY });
-  };
-
-  const handleLogout = async (accountId: string) => {
-    setCtxMenu(null);
-    const account = accountsById[accountId];
-    if (account) {
-      trackEvent('tauri_browser_click', {
-        surface: 'chat_right_sidebar',
-        action: 'disconnect_account',
-        provider: account.provider,
-        account_status: account.status ?? 'unknown',
-      });
-    }
-    try {
-      await purgeWebviewAccount(accountId);
-    } catch {
-      // Purge failures are already logged by the service; still drop the
-      // account from the UI so the user isn't stuck with a zombie icon.
-    }
-    dispatch(removeAccount({ accountId }));
-  };
-
-  // Close the context menu on Escape or any outside click.
-  useEffect(() => {
-    if (!ctxMenu) return;
-    const close = () => setCtxMenu(null);
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') close();
-    };
-    window.addEventListener('mousedown', close);
-    window.addEventListener('keydown', onKey);
-    return () => {
-      window.removeEventListener('mousedown', close);
-      window.removeEventListener('keydown', onKey);
-    };
-  }, [ctxMenu]);
-
   return (
     <div
       // `h-full` makes this page fill the shell's content box edge-to-edge.
       className="relative flex h-full overflow-hidden"
       data-testid="accounts-page"
       data-analytics-id="chat-right-sidebar">
-      {/* App rail — projected into the root sidebar's dynamic region as a compact
-          horizontal row above the thread search (order-0 sits above the thread
-          list, which Conversations projects as order-1). */}
-      <SidebarContent>
-        <div
-          data-testid="accounts-app-rail"
-          data-analytics-id="chat-app-rail"
-          className="scrollbar-hide order-0 flex flex-none items-center gap-1.5 overflow-x-auto overflow-y-hidden border-b border-stone-100 px-2 py-2 dark:border-neutral-800">
-          <RailButton
-            active={isAgentSelected}
-            onClick={selectAgent}
-            tooltip={t('accounts.agent')}
-            analyticsId="chat-app-rail-agent">
-            <AgentIcon className="h-5 w-5 rounded-md bg-white dark:bg-neutral-200" />
-          </RailButton>
-
-          {accounts.map(acct => (
-            <RailButton
-              key={acct.id}
-              active={acct.id === selectedId}
-              onClick={() => selectAccount(acct.id)}
-              onContextMenu={e => openContextMenu(acct.id, e)}
-              tooltip={acct.label}
-              analyticsId={`chat-app-rail-account-${acct.provider}`}
-              badge={unreadByAccount[acct.id]}>
-              <ProviderIcon provider={acct.provider} className="h-5 w-5 rounded" />
-            </RailButton>
-          ))}
-
-          <button
-            type="button"
-            onClick={() => {
-              trackEvent('tauri_browser_click', {
-                surface: 'chat_app_rail',
-                action: 'open_add_account',
-                provider: 'none',
-              });
-              setAddOpen(true);
-            }}
-            data-analytics-id="chat-app-rail-add-account"
-            data-testid="accounts-add-button"
-            className="group relative flex h-9 w-9 flex-none items-center justify-center rounded-xl border border-dashed border-stone-300 text-stone-400 hover:bg-stone-50 hover:text-stone-600 dark:border-neutral-700 dark:text-neutral-500 dark:hover:bg-neutral-800/60 dark:hover:text-neutral-300"
-            aria-label={t('accounts.addAccount')}
-            title={t('accounts.addAccount')}>
-            <svg className="h-4 w-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M12 4v16m8-8H4"
-              />
-            </svg>
-          </button>
-        </div>
-      </SidebarContent>
-
       {/* "Talk to Tiny" face-mode toggle — hidden (kept for potential re-enable). */}
 
       {/* Main pane. In face mode (agent selected) it's a horizontal split with
@@ -431,36 +207,6 @@ const Accounts = () => {
             </div>
           )}
         </main>
-      )}
-
-      <AddAccountModal
-        open={addOpen}
-        onClose={() => setAddOpen(false)}
-        onPick={handlePickProvider}
-        connectedProviders={connectedProviders}
-      />
-
-      {ctxMenu && (
-        <div
-          className="fixed z-50 min-w-[140px] rounded-lg border border-stone-200 dark:border-neutral-800 bg-white dark:bg-neutral-900 py-1 shadow-strong"
-          style={{ left: ctxMenu.x, top: ctxMenu.y }}
-          onMouseDown={e => e.stopPropagation()}>
-          <button
-            type="button"
-            data-analytics-id="chat-right-sidebar-disconnect-account"
-            onClick={() => void handleLogout(ctxMenu.accountId)}
-            className="flex w-full items-center gap-2 px-3 py-1.5 text-left text-sm text-coral-600 hover:bg-stone-100 dark:hover:bg-neutral-800/60">
-            <svg className="h-4 w-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"
-              />
-            </svg>
-            {t('accounts.disconnect')}
-          </button>
-        </div>
       )}
     </div>
   );
