+++ app/src/store/__tests__/accountsSlice.rehydrate.test.ts
@@ -33,6 +33,7 @@ function seedState(accounts: Account[]): AccountsState {
     messages: {},
     unread: {},
     logs: {},
+    overlayOpen: false,
   };
   for (const acct of accounts) {
     state.accounts[acct.id] = acct;
